refactor: 기존 컴포넌트들 모두 제거 및 global style 적용
