--- src/core/Component.js
@@ -1,40 +1,0 @@
-export default class Component {
-  $target;
-  props;
-  state;
-
-  constructor($target, props) {
-    this.$target = $target;
-    this.props = props;
-    this.setup();
-    this.render();
-    this.setEvent();
-  }
-
-  setup() {}
-
-  mounted() {}
-
-  template() {
-    return "";
-  }
-
-  render() {
-    this.$target.innerHTML = this.template();
-    this.mounted();
-  }
-
-  setEvent() {}
-
-  setState(newState) {
-    this.state = { ...this.state, ...newState };
-    this.render();
-  }
-
-  addEvent(eventType, selector, callback) {
-    this.$target.addEventListener(eventType, (event) => {
-      if (!event.target.closest(selector)) return;
-      callback(event);
-    });
-  }
-}
