+++ src/index.js
@@ -6,5 +6,5 @@ import "./images/category-japanese.png";
 import "./images/category-korean.png";
 import "./images/category-western.png";
 
-import "../templates/style.css";
+import "./global.css";
 import "./App";
